+++ CHANGES.rst
@@ -6,6 +6,8 @@ Unreleased
     used without an explicit value. :issue:`3084`
 -   Hide ``Sentinel.UNSET`` values as ``None`` when using ``lookup_default()``.
     :issue:`3136` :pr:`3199` :pr:`3202` :pr:`3209` :pr:`3212` :pr:`3224`
+-   Prevent ``_NamedTextIOWrapper`` from closing streams owned by ``StreamMixer``.
+    :issue:`824` :issue:`2993` :pr:`3139`
 
 Version 8.3.1
 --------------